refactor: 훅 파일 이동
